Huge changes to mostly code structure itself. Added a command for users to change what their interactions with an ender chest open along with better config options that allow any ender chest to be enabled or disabled.
To do so, many save file names have been changed and so will not pick up save files from older versions.

Some testing has occured, but not extensive testing and so due to the complexity some bugs maybe found. However, it seems very usable right now and so will be marked as release 2.0.0